[dl-cifar] Optimise initImage using static cache
The initImage function is called hundreds of times,
repeating the same computations on the host each time.
The redundant computation introduces large host overhead
in the measured computation times and skews the comparison
between offload programming models.

Speed up the initImage function by using a static cache
in order to minimise the host overheads in the benchmarked
computation.